store pointers to the objects.
it seems getting a value from a map inside for loop or select statement
returns us a clone of that object, not the actual object itself. so, our
changes don't get persisted. storing pointers of the query and query
meta objects' pointers instead of full objects has resolved the issue.